+++ test/ie.Tests.ps1
@@ -36,9 +36,10 @@ Describe 'ie tests' {
   Context "PlatformNeutral" {
 
     It 'Passes script-block-based PowerShell calls through as-is' {
-      ('pwsh', ('pwsh.exe', 'powershell.exe'))[$IsWindows].ForEach({
-        ie pwsh -noprofile -c { "[$args]" } -args one, two | Should -Be '[one two]'
-      })
+      foreach ($cli in ('pwsh', ('pwsh.exe', 'powershell.exe'))[$IsWindows]) {
+        if (-not (Get-Command -ErrorAction Ignore -CommandType Application $cli)) { Write-Warning "Ignoring unavailability of '$cli'."; continue } # Ignore nonexistence of pwsh.exe on Windows, for testing on the WinPS v3 and v4 VMs
+        ie $cli -noprofile -c { "[$args]" } -args one, two | Should -Be '[one two]'
+      }
     }
 
     It 'Only permits calls to external executables' {
@@ -61,7 +62,7 @@ Describe 'ie tests' {
   
       # Note: Avoid arguments with embedded newlines, because dbea -Raw
       #       doesn't support them due to line-by-line output.
-      $exeArgs = '', 'a&b', '3 " of snow', 'Nat "King" Cole', 'c:\temp 1\', 'a b\\', 'a \" b', 'a \"b c\" d', 'a"b', 'ab\'
+      $exeArgs = '', 'a&b', '3 " of snow', 'Nat "King & I" Cole', 'c:\temp 1\', 'a b\\', 'a \" b', 'a \"b c\" d', 'a"b', 'ab\'
   
       $result = dbea -Raw -UseIe -- $exeArgs
       assert-ExpectedResult $exeArgs $result
@@ -84,7 +85,7 @@ Describe 'ie tests' {
   
       foreach ($exe in $ohtCmds.Keys) {
         if (Get-Command -ea Ignore -Type Application $exe) {
-          "Testing with $exe...." | Write-Verbose -vb
+          "Testing with installed CLI '$exe'...." | Write-Verbose -vb
           $expected = if ($exe -eq 'Rscript') { '[1] "hi there"' } else { 'hi there' }
           & $ohtCmds[$exe] | Should -BeExactly $expected
         }
@@ -107,72 +108,42 @@ Describe 'ie tests' {
       assert-ExpectedResult $expected (dbea -UseIe -UseWSH -Raw $exeArgs)
     }
 
-    It 'Handles batch-file quoting needs with space-less arguments' {
-  
-      # cmd.exe metachars. (in addition to ") that must trigger "..." enclosure:
-      #   & | < > ^ , ;
-      # Note: `=` too requires "..." enclosure in batch-file argument parsing, but it conflicts
-      #       with keeping the LHS of tokens such as `FOO=bar` *unquoted* to support misexec-style CLIs,
-      #       and we give precedence to the latter rule.
-      $exeArgs =
-      'a"b', 'a&b', 'a|b', 'a<b', 'a>b', 'a^b', 'a,b', 'a;b', 'a=b', 'the last one'
-
-      # Note: Batch files always echo arguments exactly as quoted.
-      #       Note the - unavoidable - spltting of 'a=b' into 'a' and 'b'
-      $expected =
-      '"a""b"', '"a&b"', '"a|b"', '"a<b"', '"a>b"', '"a^b"', '"a,b"', '"a;b"', 'a', 'b', '"the last one"' 
-  
-      $actual = (dbea -UseIe -UseBatchFile -Raw -- $exeArgs) -split '\r\n'
-      assert-ExpectedResult $expected $actual        
-  
-    }
-
     It 'Handles partial-quoting needs of CLIs such as msiexec.exe' {
   
       # Arguments of the following form must be placed with *partial double-quoting*,
       # *around the value only* on the command line (e.g., `foo="bar none"`):
-      $argsToPartiallyQuote = if ($PSVersionTable.PSVersion.Major -le 4) {
-        # !! CAVEAT: In WinPS v3 and v4 only, *a partially quoted value that contains spaces*, such as `foo="bar none"`,
-        # !! causes the engine to still enclose the entire argument in "...", which cannot be helped.
-        # !! Only a space-less value that needs quoting is handled correctly (which may never occur in practice).
-        Write-Warning "Partial quoting of msiexec-style arguments with spaces not supported in v3 and v4, skipping test."
-        'foo=bar"none'
-      }
-      else {
-        # !! In PS Core, verbatim `foo=bar"none` would only become `'foo="bar""none"` if the executable were msiexec.exe, msdeploy.exe, or cmdkey.exe; 
-        # !! otherwise - as in these tests - \"-escaping kicks in and `foo=bar"none` becomes `foo=bar\"none` instead - without syntactic double-quoting around the value.
-        'foo=bar none', '/foo:bar none', '-foo:bar none', 'foo=bar "stuff" none' + ('foo=bar"none', @())[$IsCoreCLR]
-      }
-      
+      $argsToPartiallyQuote = 'foo=bar none', '/foo:bar none', '-foo:bar none', 'foo=bar "stuff" none','foo=bar"none'
+
       $argsToPartiallyQuote | ForEach-Object {
   
+        # Set the test-override environment variable to force `ie` to use ""-escaping, as
+        # if msiexec.exe, msdeploy.exe or cmdkey.exe were being called, the hard-coded
+        # msiexec-like executables that 
+        $env:__ie_doubledquotes = 1
+
         $exeArgs = $_, 'a " b'  # Add a second argument, as a control.
 
-        # The value part be selectively quoted.
-        # ""-escaping must also be triggered (it's what msiexec requires).
-        # !! In WinPS we use ""-escaping by default anyway, to work around legacy bugs, 
-        # !! but in *PS Core* we default to \"-escaping, because it is the safer choice; as a nod to 
-        # !! these high-profile CLIs, we have *hard-coded exceptions for 'msiexece' and 'msdeploy' there.
-        # !! We therefore can't test this aspect on PS Core. A manual way to do it is to create (temporary)
-        # !! copies of de.exe, name them 'msiexec.exe' and 'msdeploy.exe', then call them with `ie`; e.g.:
-        # !!
-        # !!   'msiexec.exe', 'msdeploy.exe' | % { cpi -ea stop (gcm de.exe).Path "./$_"; ie "./$_" 'foo=bar none', '/foo:bar none', '-foo:bar none', 'foo=bar "stuff" none', 'foo=bar"none'; ri "./$_" }
-        # !!  
-        $partiallyQuotedArg = (($_ -replace '"', ('""', '\"')[$IsCoreCLR]) -replace '(?<=[:=]).+$', '"$&"') # turn `'foo=bar none` into `foo="bar none"`, ...
-        $expectedRawCmdLine = '{0} {1}' -f $partiallyQuotedArg, ('"a "" b"', '"a \" b"')[$IsCoreCLR]
-
-        # Run dbea and extract the raw command line from the output (last non-blank line.)
-        $rawCmdLine = ((dbea -UseIe -- $exeArgs) -notmatch '^\s*$')[-1].Trim()
-        $rawCmdLine | Should -BeExactly $expectedRawCmdLine
-
-        # --- Should also work when a *batch file* is (first) invoked, in which case ""-escaping must be used even in PS Core.
-        # See above.
-        $partiallyQuotedArg = (($_ -replace '"', '""') -replace '(?<=[:=]).+$', '"$&"')
-        $expectedRawCmdLine = '{0} {1}' -f $partiallyQuotedArg, '"a "" b"'
-
-        # Run dbea *via a wrapper batch file* and extract the raw command line from the output (last non-blank line.)
-        $rawCmdLine = ((dbea -UseWrapperBatchFile -UseIe -- $exeArgs) -notmatch '^\s*$')[-1].Trim()
-        $rawCmdLine | Should -BeExactly $expectedRawCmdLine
+        try {
+          # The value part must be selectively quoted.
+          $partiallyQuotedArg = ($_ -replace '"', '""' -replace '(?<=[:=]).+$', '"$&"') # turn `'foo=bar none` into `foo="bar none"`, ...
+          $expectedRawCmdLine = '{0} {1}' -f $partiallyQuotedArg, '"a "" b"'
+  
+          # Run dbea and extract the raw command line from the output (last non-blank line.)
+          $rawCmdLine = ((dbea -UseIe -- $exeArgs) -notmatch '^\s*$')[-1].Trim()
+          $rawCmdLine | Should -BeExactly $expectedRawCmdLine
+  
+          # --- Should also work when a *batch file* is (first) invoked, in which case ""-escaping must be used even in PS Core.
+          # See above.
+          $partiallyQuotedArg = (($_ -replace '"', '""') -replace '(?<=[:=]).+$', '"$&"')
+          $expectedRawCmdLine = '{0} {1}' -f $partiallyQuotedArg, '"a "" b"'
+  
+          # Run dbea *via a wrapper batch file* and extract the raw command line from the output (last non-blank line.)
+          $rawCmdLine = ((dbea -UseWrapperBatchFile -UseIe -- $exeArgs) -notmatch '^\s*$')[-1].Trim()
+          $rawCmdLine | Should -BeExactly $expectedRawCmdLine
+        }
+        finally {
+          $env:__ie_doubledquotes = $null
+        }
         
       }  
 
@@ -182,7 +153,7 @@ Describe 'ie tests' {
     It 'Handles a *single-argument* command-line cmd.exe /c call correctly.' {
   
       $expected = 'Ready to move on [Y,N]?Y'
-      $choicePath = (Get-Command -ea Stop choice.exe).Source
+      $choicePath = (Get-Command -ea Stop choice.exe).Path # !! Use .Path, not .Source - the latter doesn't work in WinPS v3.
 
       # Command line passed as single argument with embedded quoting.
       ie cmd.exe /c (' "{0}" /d Y /t 0 /m "Ready to move on" ' -f $choicePath) | Should -BeExactly $expected
@@ -192,7 +163,7 @@ Describe 'ie tests' {
     It 'Handles a *multi-argument* command-line cmd.exe /c call correctly.' {
   
       $expected = 'Ready to move on [Y,N]?Y'
-      $choicePath = (Get-Command -ea Stop choice.exe).Source
+      $choicePath = (Get-Command -ea Stop choice.exe).Path # !! Use .Path, not .Source - the latter doesn't work in WinPS v3.
 
       # We want to use an executable path *with spaces*, so that PowerShell
       # "..."-encloses it, as that is the real test for whether `ie` correctly
@@ -209,17 +180,66 @@ Describe 'ie tests' {
         # part of the invocation should behave the same between editions.
       }
 
-      # Similiar command line as above, but passed as multiple arguments,
-      # with the executable path containing spaces, if feasible.
-      ie cmd.exe /c $choicePath /d Y /t 0 /m "Ready to move on" | Should -BeExactly $expected
-
-      if ($IsCoreCLR)  {
-        Remove-Item -LiteralPath $tmpSymlinkWithSpaces
+      try {
+        # Similiar command line as above, but passed as multiple arguments,
+        # with the executable path containing spaces, if feasible.
+        ie cmd.exe /c $choicePath /d Y /t 0 /m "Ready to move on" | Should -BeExactly $expected
+      }
+      finally {
+        if ($IsCoreCLR)  {
+          Remove-Item -LiteralPath $tmpSymlinkWithSpaces
+        }
       }
   
     }
 
-    It 'Robust exit-code reporting works' {
+    It 'Passing arguments with cmd.exe metacharacters to batch file works' {
+
+      $tempBatFile = Join-Path (Get-Item TestDrive:/).FullName 'tmp.cmd'
+      $exeArgs = 'foo1&bar1', 'foo2|bar2', 'foo3^bar3', 'foo4<bar4', 'foo5>bar5', 'foo6,bar6', 'foo7;bar7', 'foo8=bar8', 'more "stuff"', 'last'
+      # Note how all space-less args with metachars. must end up double-quoted, except foo8=bar8, due to the misexec-like exception.
+      $expected = '"foo1&bar1" "foo2|bar2" "foo3^bar3" "foo4<bar4" "foo5>bar5" "foo6,bar6" "foo7;bar7" foo8=bar8 "more ""stuff""" last'
+
+      try {
+
+          '@echo %*' | Set-Content -LiteralPath $tempBatFile
+
+          $output = ie $tempBatFile $exeArgs 2>&1           
+          $output | Should -Be $expected
+          $LASTEXITCODE | Should -Be 0
+
+      }
+      finally {
+        Remove-Item $tempBatFile
+      }
+
+    }
+
+
+    It 'Invoking batch files with quoted paths works' {
+
+      $tempBatFile = Join-Path (Get-Item TestDrive:/).FullName 'tmp 1.cmd'
+      $exeArgs = 'foo1&bar1', 'foo2|bar2', 'foo3^bar3', 'foo4<bar4', 'foo5>bar5', 'foo6,bar6', 'foo7;bar7', 'foo8=bar8', 'more "stuff"', 'last'
+      # Note how all space-less args with metachars. must end up double-quoted, except foo8=bar8, due to the misexec-like exception.
+      $expected = '"foo1&bar1" "foo2|bar2" "foo3^bar3" "foo4<bar4" "foo5>bar5" "foo6,bar6" "foo7;bar7" foo8=bar8 "more ""stuff""" last'
+
+      try {
+
+          '@echo %*' | Set-Content -LiteralPath $tempBatFile
+
+          $output = ie $tempBatFile $exeArgs 2>&1 
+          $output | Should -Be $expected
+          $LASTEXITCODE | Should -Be 0
+
+      }
+      finally {
+        Remove-Item $tempBatFile
+      }
+
+    }
+
+
+    It 'Robust batch-file exit-code reporting works' {
 
       # For added testing:
       #  * use a batch-file name with spaces
@@ -227,19 +247,29 @@ Describe 'ie tests' {
       $tempBatFile = Join-Path (Get-Item TestDrive:/).FullName 'tmp 1.cmd'
       $dummyArgs = 'foo&bar', 'unrelated "stuff"'
 
-      # Create a temporary batch file that uses exit /b *without* an explicit
-      # exit code, which should pass the failing command's exit code (error level) through.
-      # This only happens when the batch file is called via `cmd /c ... & exit `
-      # See https://stackoverflow.com/q/66975883/45375
-      '@echo off & whoami -nosuch 2>NUL || exit /b' | Set-Content -LiteralPath $tempBatFile
+      try {
+
+          # Create a temporary batch file that uses exit /b *without* an explicit
+          # exit code, which should pass the failing command's exit code (error level) through.
+          # This only happens when the batch file is called via `cmd /c ... & exit `
+          # See https://stackoverflow.com/q/66975883/45375
+          '@echo off & whoami -nosuch 2>NUL || exit /b' | Set-Content -LiteralPath $tempBatFile
+
+          $output = ie $tempBatFile $dummyArgs 2>&1  
+          $output | Should -BeNullOrEmpty
+          $LASTEXITCODE | Should -Be 1
+
+          # As a control: Rewrite the batch file to succeed.
+          '@echo off & whoami & exit /b' | Set-Content -LiteralPath $tempBatFile
+          $output = ie $tempBatFile $dummyArgs 2>&1  
+          $output | Should -Be (whoami)
+          $LASTEXITCODE | Should -Be 0
+
+      }
+      finally {
+        Remove-Item $tempBatFile
+      }
       
-      $output = ie $tempBatFile $dummyArgs 2>&1
-
-      Remove-Item $tempBatFile
-
-      $output | Should -BeNullOrEmpty
-      $LASTEXITCODE | Should -Be 1
-
     }
 
   }